pyproject.toml
[project]
name = "onyxchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiomysql>=0.2.0",
    "asyncio>=4.0.0",
    "python-dotenv>=1.1.1",
]
